feat(YouTube Music - AMOLED black theme): Add `AMOLED black theme` patch
