+++ extensions/shared/src/main/java/app/revanced/extension/shared/ThemeHelper.java
